# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentio SMTP API",
    "description": "Multi-tenant SMTP platform API",
    "contact": {
      "name": "Truespar Team"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/analytics/bounces": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "bounces",
        "parameters": [
          {
            "name": "from",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_BounceStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/analytics/engagement": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "engagement",
        "parameters": [
          {
            "name": "from",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_EngagementStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/analytics/overview": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "overview",
        "parameters": [
          {
            "name": "from",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OverviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/messages/{id}/attachments": {
      "get": {
        "tags": [
          "Messages"
        ],
        "operationId": "list_attachments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Vec_AttachmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/messages/{message_id}/attachments/{attachment_id}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "operationId": "download_attachment",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachmentDisposition": {
        "type": "string",
        "enum": [
          "attachment",
          "inline"
        ]
      },
      "AttachmentId": {
        "type": "string",
        "format": "uuid"
      },
...
      "DataResponse_BounceStatsResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "hard",
              "soft",
              "block",
              "total"
            ],
            "properties": {
              "block": {
                "type": "integer",
                "format": "int64"
              },
              "hard": {
                "type": "integer",
                "format": "int64"
              },
              "soft": {
                "type": "integer",
                "format": "int64"
              },
              "total": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      },
...
      "DataResponse_EngagementStatsResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "opens",
              "clicks",
              "unsubscribes"
            ],
            "properties": {
              "clicks": {
                "type": "integer",
                "format": "int64"
              },
              "opens": {
                "type": "integer",
                "format": "int64"
              },
              "unsubscribes": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      },
...
      "DataResponse_OverviewResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "total_sent",
              "delivered",
              "bounced",
              "deferred",
              "dropped",
              "delivery_rate",
              "bounce_rate"
            ],
            "properties": {
              "bounce_rate": {
                "type": "number",
                "format": "double"
              },
              "bounced": {
                "type": "integer",
                "format": "int64"
              },
              "deferred": {
                "type": "integer",
                "format": "int64"
              },
              "delivered": {
                "type": "integer",
                "format": "int64"
              },
              "delivery_rate": {
                "type": "number",
                "format": "double"
              },
              "dropped": {
                "type": "integer",
                "format": "int64"
              },
              "total_sent": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      },
...
      "DataResponse_Vec_AttachmentResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "message_id",
                "filename",
                "content_type",
                "size",
                "disposition",
                "scan_status",
                "created_at"
              ],
              "properties": {
                "content_id": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "content_type": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "disposition": {
                  "$ref": "#/components/schemas/AttachmentDisposition"
                },
                "filename": {
                  "type": "string"
                },
                "id": {
                  "$ref": "#/components/schemas/AttachmentId"
                },
                "message_id": {
                  "$ref": "#/components/schemas/MessageId"
                },
                "scan_result": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "scan_status": {
                  "$ref": "#/components/schemas/ScanStatus"
                },
                "size": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "MessageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ScanStatus": {
        "type": "string",
        "enum": [
          "pending",
          "clean",
          "infected",
          "error"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
...
}